pyproject.toml
[tool.poetry]
name = "sas-backend"
version = "0.1.0"
description = ""
authors = ["Vladimir Pasquier <1112241+vpasquier@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.4"
djangorestframework = "^3.15.1"
Markdown = "^3.6"
djoser = "^2.2.2"
django-filter = "^24.2"
pillow = "^10.3.0"
django-cors-headers = "^4.3.1"
django-extensions = "^3.2.3"
pydotplus = "^2.0.2"
faker = "^25.2.0"
django-factory-boy = "^1.0.0"
factory-boy = "^3.3.0"
pytest-django = "^4.8.0"

[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.1"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
